+++ attributes/ioncube.rb
@@ -1,8 +1,8 @@
 # # # # # #
 # ioncube #
 # # # # # #
-default['php']['ioncube']['filename'] = 'ioncube_loaders_lin_x86-64'
+default['php']['ioncube']['home'] = '/usr/lib/php/ioncube'
+default['php']['ioncube']['filename'] = 'ioncube_loaders_lin_x86-64.tar.gz'
 default['php']['ioncube']['checksum'] = '21c37a244224420a76fc4705c43d221e6d32ff592ec633eee4f6bb6a394c6402'
-default['php']['ioncube']['download_url'] = "http://downloads3.ioncube.com/loader_downloads/#{node['php']['ioncube']['filename']}.tar.gz"
-default['php']['ioncube']['home'] = '/usr/lib/php5/ioncube'
+default['php']['ioncube']['download_url'] = "http://downloads3.ioncube.com/loader_downloads/#{node['php']['ioncube']['filename']}"
 default['php']['ioncube']['version'] = 'ioncube_loader_lin_5.5'